# 💲[Native Checkout] Pledge Screen Separators