# 📲 What

Adds separators to the description section and summary section.

# 🤔 Why

To visually distinguish between various sections in the pledge screen.

# 🛠 How

This PR adds two separator views: `descriptionSectionSeparator` and `summarySectionSeparator`, and also adds logical view array groupings that define which views should belong to each section. This makes it easier to reason about which views are logically grouped, and will come in handy later when we want to conditionally hide/show different sections. For example, if one wishes to hide the `pledgeSummarySection`, instead of creating two different output signals to hide the `summarySectionSeparator` and the `pledgeSummaryViewController.view`, one could do some thing like the following:

```
self.viewModel.outputs.hideSummarySection
  .observeForUI()
  .observeValues { isHidden in
      _ = self.summarySectionViews
        ||> \.isHidden .~ isHidden
  }
```

# 👀 See

Trello, screenshots, external resources?

| Before 🐛 | After 🦋 |
| --- | --- |
|![Simulator Screen Shot - iPhone 5s - 2019-07-18 at 16 37 26](https://user-images.githubusercontent.com/3156796/61554722-97e20b00-aa2b-11e9-8220-afa7e8447525.png)|![Simulator Screen Shot - iPhone 5s - 2019-07-19 at 13 22 23](https://user-images.githubusercontent.com/3156796/61554739-aa5c4480-aa2b-11e9-9d4f-11e17625f9ff.png)|

See also:

| Logged Out, No Shipping Rules | Logged Out, Shipping Rules |
| --- | --- |
| ![Simulator Screen Shot - iPhone X - 2019-07-19 at 13 26 22](https://user-images.githubusercontent.com/3156796/61554845-ea232c00-aa2b-11e9-94d4-83c05bd58e71.png) | ![Simulator Screen Shot - iPhone X - 2019-07-19 at 13 26 32](https://user-images.githubusercontent.com/3156796/61554854-f3ac9400-aa2b-11e9-803e-4cd044f4d6d8.png) |



# ♿️ Accessibility 

N/A

# 🏎 Performance

N/A

# ✅ Acceptance criteria

**Logged in**

1. Log in to an account with the native checkout feature flag enabled.
2. Navigate to any project.
3. Select a reward without shipping enabled.
4. **You should the the pledge screen with two separators, without the shipping row, and they should match the designs.**
5. Navigate back to rewards and select a reward with shipping enabled.
6. **You should see the pledge screen with two separators and the shipping row, and they should match the designs.**

**Logged out**

1. When you are logged out, navigate to the feature flag tools and toggle the "Native checkout" feature flag to **ON**.
2. Navigate to any project and select a reward without shipping enabled.
3. **You should see the pledge screen with two separators and the continue button, and they should match the designs.**
4. Navigate back to rewards and select a reward with shipping enabled.
5. **You should see the pledge screen with two separators and the shipping row, and the continue row, and they should match the designs.**