@ifbarrera 2019-07-22
> Looking good, couple comments and then I'm just wondering:
> 
> 1. Do we get much benefit in overall clarity by making these computed arrays `var`s and not just constructing them once in the function where we use them? I guess my question is - will we ever refer to them in multiple places?
> 2. I know we chatted in passing about the fact that separators may need to be hidden/shown later on when this view's functionality grows. Remind me how that will work?
> 3. There seems to be a constraint warning: <img alt="Untitled" width="958" src="https://user-images.githubusercontent.com/3735375/61655338-6451eb80-ac8c-11e9-905d-beb507ce42bd.png"> Changing the priority of the constraint on line `142` of `AmountInputView.swift` to `.defaultHigh` appears to fix this.

1. The main reason for storing a reference to these sections at the VC level was so we can easily hide/show groupings of views (ie. sections) like in the example described in the *How* section of the PR description.

2. See the example in the *How* section. The separators will always be hidden/shown as part of their section group.

3. Would you mind drafting a small card for us to fix this constraint warning? 🙏