@ifbarrera 2019-07-25
![](https://github.trello.services/images/mini-trello-icon.png) [[Checkout] Section separator](https://trello.com/c/SjubmRvB/1300-checkout-section-separator)